Rename point -> geo
